+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 ## [Unreleased]
+### Added
+- Validator `#type_valid?` for checking the initial type of a model
+- `.caching` option for generating the iVar and reader on the class
 
 ## [0.2.0] - 2023-12-20
-
+### Changed
 - Attempt to clean up all code
 
 ## [0.1.0] - 2023-12-19
-
-- Gem created release
+### Added
+- Initial gem creation
 
 [Unreleased]: https://github.com/site-prism/testingrecord/compare/v0.2...HEAD
 [0.2.0]:      https://github.com/site-prism/testingrecord/compare/0.1...v0.2